+++ TDEngine2/plugins/D3D11GraphicsContext/source/CD3D11ShaderCompiler.cpp
@@ -442,7 +442,9 @@ namespace TDEngine2
 			/// skip template parameters
 			if (E_SHADER_RESOURCE_TYPE::SRT_STRUCTURED_BUFFER == currType || E_SHADER_RESOURCE_TYPE::SRT_RW_STRUCTURED_BUFFER == currType ||
 				E_SHADER_RESOURCE_TYPE::SRT_RAW_BUFFER == currType || E_SHADER_RESOURCE_TYPE::SRT_RW_RAW_BUFFER == currType ||
-				E_SHADER_RESOURCE_TYPE::SRT_RW_IMAGE2D == currType || E_SHADER_RESOURCE_TYPE::SRT_RW_IMAGE3D == currType)
+				E_SHADER_RESOURCE_TYPE::SRT_RW_IMAGE2D == currType || E_SHADER_RESOURCE_TYPE::SRT_RW_IMAGE3D == currType ||
+				(E_SHADER_RESOURCE_TYPE::SRT_TEXTURE2D == currType && tokenizer.Peek(1) == "<") ||
+				(E_SHADER_RESOURCE_TYPE::SRT_TEXTURE3D == currType && tokenizer.Peek(1) == "<"))
 			{
 				while (currToken != ">")
 				{